chore: removing circular dependencies
